# backend/app/utils/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "FitHub - API",
        "description": "Welcome to FitHub api's documentation. Here you will find all the routes at disposal including details and their HTTP methods."
    },
...
    "paths": {
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Fetch all users & their public informations.",
                "description": "Returns an array of objects with users' informations.",
                "operationid": "findAll",
                "responses": {
                    "200": {
                        "description": " Returns a JSON object with users' data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Users cannot be found."
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new user account.",
                "description": "Create a new user account based on the data retrieved.",
                "operationId": "createOne",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created !"
                    },
                    "400": {
                        "description": "Missing required fields."
                    },
                    "409": {
                        "description": "Nickname, email or phone already linked to a user."
                    }
                }
            }
        },
...
        "/category-activity" : {
            "get": {
                "tags": [
                    "Category activity"
                ],
                "summary": "Fetch all categories of activities.",
                "description": "Fetch all categories of activities",
                "operationId": "findAll",
                "responses": {
                    "200": {
                        "description": " Returns a JSON array object with all categories.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categories not found."
                    }
                }
            },
            "post": {
                "tags": [
                    "Category activity"
                ],
                "summary": "Create a new category.",
                "description": "Create a new activities' category.",
                "operationId": "createOne",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref":"#/components/schemas/Category activity"
                            }
                        }
                    }
                },
                "responses": {
                    "200" : {
                        "description": "Category created !"
                    },
                    "400": {
                        "description": "Label is required."
                    },
                    "409": {
                        "description": "Category already exists."
                    }
                }
            }
        },
...
        "/challenge/:challengeId": {
            "get": {
                "tags": [
                    "Challenge"
                ],
                "summary": "Fetch one challenge.",
                "description": "Fetch one challenge.",
                "operationId": "findOne",
                "parameters": [
                    {
                        "name": "challengeId",
                        "required": "true",
                        "in": "path",
                        "description": "ID of category to return."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a JSON object with the challenge data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref":"#/components/schemas/Challenge"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Challenge cannot be found."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Challenge"
                ],
                "summary": "Update a challenge.",
                "description": "Update a challenge.",
                "operationid": "updateOne",
                "parameters": [
                    {
                        "name": "challengeId",
                        "required": "true",
                        "in": "path",
                        "description": "ID of challenge to update."
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Challenge"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Challenge updated !"
                    },
                    "404": {
                        "description": "Challenge cannot be found."
                    },
                    "409": {
                        "description": "Challenge already exists."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Challenge"
                ],
                "summary": "Delete a challenge",
                "description": "Delete a challenge",
                "operationId": "deleteOne",
                "parameters": [
                      {
                        "name": "challengeId",
                        "required": "true",
                        "in": "path",
                        "description": "ID of challenge to delete."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge deleted !"
                    },
                    "404": {
                        "description": "Challenge cannot be found."
                    }
                }
            }
        },
        "challenge/user": {
            "post": {
                "tags": [
                    "Challenge"
                ],
                "summary": "Assign a random challenge to a user",
                "description": "Assign a random challenge to a user",
                "operationId": "assignChallenge",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref":"#/components/schemas/Assign challenge"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Challenge assigned !"
                    },
                    "404": {
                        "description": "Challenge or user cannot be found."
                    },
                    "409": {
                        "description": "User already receveived a challenge today."
                    }
                }
            }
        },
        "/challenge/user/:userId": {
            "patch": {
                "tags": [
                    "Challenge"
                ],
                "summary": "Checks or unchecks a challenge completion.",
                "description": "Checks or unchecks a challenge completion",
                "operationId": "challengeChecker",
                "parameters": [
                    {
                        "name": "userId",
                        "required": "true",
                        "description": "ID of user to update their challenge status.",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge completion updated !"
                    },
                    "400":{
                        "description": "Only current day's challenge can be validated."
                    },
                    "404": {
                        "description": "Challenge or user cannot be found. Otherwise challenge was not assigned."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "firstname": {
                        "type": "string",
                        "example": "John",
                        "required": "true"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Doe",
                        "required": "true"
                    },
                    "nickname": {
                        "type": "string",
                        "example": "John_Doe_73",
                        "required": "true"
                    },
                    "phone": {
                        "type": "string",
                        "example": "0123456789",
                        "required": "false",
                        "default": "null"
                    },
                    "role": {
                        "type": "string",
                        "example": "user",
                        "enum": [
                            "user",
                            "admin"
                        ],
                        "default": "user"
                    },
                    "login_streak": {
                        "type": "integer",
                        "example": 10,
                        "default": 0
                    },
                    "password": {
                        "type": "string",
                        "example": "$2b$10$wyzLD2T8Jxpb2Fp4r7cgWOapKwfTGkLGMKq1VVwsc2r4LTNUDXIhq",
                        "required": "true"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe73@mymail.com",
                        "required": "true"
                    },
                    "weight": {
                        "type": "integer",
                        "example": 64,
                        "required": "true"
                    },
                    "age": {
                        "type": "integer",
                        "example": 25,
                        "required": "true"
                    },
                    "gender": {
                        "type": "string",
                        "example": "homme",
                        "enum": [
                            "femme",
                            "homme",
                            "non-spécifié"
                        ],
                        "default": "non-spécifié"
                    },
                    "challenge_id": {
                        "type": "integer",
                        "example": 2,
                        "default": "null"
                    },
                    "xp": {
                        "type": "integer",
                        "example": 478054,
                        "default": "null"
                    },
                    "profile_visibility": {
                        "type": "string",
                        "example": "privé",
                        "enum": [
                            "publique",
                            "privé"
                        ],
                        "default": "publique"
                    },
                    "image_path": {
                        "type": "string",
                        "example": "dirName/fileName123",
                        "default": "null"
                    },
                    "image_mimetype": {
                        "type": "string",
                        "example": "image/jpeg",
                        "default": "null"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "default": "2023-04-02 13:39:13.472+02"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "default": "null"
                    }
                }
            },
...
            "Challenge": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 15
                    },
                    "label": {
                        "type": "string",
                        "example": "Faire 20 pompes.",
                        "required": "true"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "default": "2023-04-02 13:39:13.472+02"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "default": "null"
                    }
                }
            },
...
        }
    }
}